requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Claude API
anthropic>=0.18.0

# HTTP requests for API calls
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Rich for CLI setup wizard
rich>=13.0.0
